feat: 프로젝트 리스트 메인페이지에서 받기
